refactor(relay): split the /files handler and its tests below the 400-LOC rule
Three findings on #302. AGENTS.md says "Keep files <~400 LOC; split as
needed", and the handler had grown to 484 LOC by absorbing access
gating, path security, filesystem work and response shaping — the exact
mix that makes a security-critical module hard to review.

Now three focused modules:
- files-paths.ts (204) — the security core: logical prefix, iterative
  decoding, control chars, `..`, realpath containment, the deny rules,
  and the delete-specific split (resolve the parent, leave the leaf).
- files-ops.ts (177) — the filesystem work: list, read, atomic write
  with .bak, delete, and the caps.
- files.ts (119) — gate, validate, dispatch, shape. Nothing else.

The 469-LOC test file is split the same way (access / paths / ops), and
scripts/test/safe-core-files.json follows so the orphan guard added
earlier stays satisfied.

Also fixed: list_dir on a regular file threw ENOTDIR and surfaced as a
500 INTERNAL_ERROR — an ordinary client mistake blamed on the server. It
is now a 400 pointing at read_file.

Verified: 74 test files / 711 tests green, typecheck clean, docs
fact-check green.
