+++ scripts/test/safe-core-files.json
@@ -23,7 +23,9 @@
   "tests/hooks/session-start.test.ts",
   "tests/http/core-proxy.test.ts",
   "tests/http/error-envelope.test.ts",
-  "tests/http/files.test.ts",
+  "tests/http/files-access.test.ts",
+  "tests/http/files-ops.test.ts",
+  "tests/http/files-paths.test.ts",
   "tests/http/health.test.ts",
   "tests/http/ws-proxy.test.ts",
   "tests/security/auth.test.ts",
